Fix: Edit blogpost functionality
The editor was not showing the initial blog post body content that we got from the server.  Also, after updating, the blog post body content was not being updated.

The issue was I was using EditorProvider and EditorContent (with useEditor) together. Now I am using just the EditorContent with useEditor.